@@ -26,20 +26,8 @@ class SessionAspect extends AbstractAspect
         'Hyperf\SocketIOServer\SocketIO::onMessage',
     ];
 
-    /**
-     * @var SessionManager
-     */
-    private $sessionManager;
-
-    /**
-     * @var ConfigInterface
-     */
-    private $config;
-
-    public function __construct(SessionManager $sessionManager, ConfigInterface $config)
+    public function __construct(private SessionManager $sessionManager, private ConfigInterface $config)
     {
-        $this->sessionManager = $sessionManager;
-        $this->config = $config;
     }
 
     public function process(ProceedingJoinPoint $proceedingJoinPoint)
